+++ src/io/tvrender.h
@@ -1,130 +1,19 @@
 #ifndef IO_TVRENDER_H
 #define IO_TVRENDER_H 1
 
+#include "tvr_internal.h"
+
 #include <stdint.h>
 
-struct tvr_msg_set_version
-{
-	int32_t version;
-};
+/* Translate versioned TVrender opcode to internal opcode */
 
-/* Vertex Buffer */
+int tvr_int_opcode(
+	uint16_t opcode,
+	int tvr_v,
+	uint16_t *int_opcode
+);
 
-struct tvr_msg_vertex_buffer_create
-{
-	int32_t vb_id;
-	uint64_t size;
-	void *data;
-};
-
-struct tvr_msg_vertex_buffer_destroy
-{
-	int32_t vb_id;
-};
-
-/* Index Buffer */
-
-struct tvr_msg_index_buffer_create
-{
-	int32_t ib_id;
-	uint64_t size;
-	void *data;
-};
-
-struct tvr_msg_index_buffer_destroy
-{
-	int32_t ib_id;
-};
-
-/* Mesh */
-
-struct tvr_msg_mesh_create
-{
-	int32_t mesh_id;
-	int32_t vb_id;
-	int32_t ib_id;
-};
-
-struct tvr_msg_mesh_bind_mat
-{
-	int32_t mesh_id;
-	int32_t mat_id;
-};
-
-struct tvr_msg_mesh_set_loc
-{
-	int32_t mesh_id;
-	float x, y, z;
-};
-
-struct tvr_msg_mesh_set_rot
-{
-	int32_t mesh_id;
-	float x, y, z;
-};
-
-struct tvr_msg_mesh_set_scale
-{
-	int32_t mesh_id;
-	float x, y, z;
-};
-
-struct tvr_msg_mesh_destroy
-{
-	int32_t mesh_id;
-};
-
-/* Texture */
-
-struct tvr_msg_texture_create
-{
-	int32_t tex_id;
-	int32_t width;
-	int32_t height;
-	int8_t channels;
-	void *data;
-};
-
-struct tvr_msg_texture_destroy
-{
-	int32_t tex_id;
-};
-
-/* Material */
-
-struct tvr_msg_material_create
-{
-	int32_t mat_id;
-};
-
-struct tvr_msg_material_bind_texture
-{
-	int32_t mat_id;
-	int32_t tex_id;
-	int8_t pass;
-};
-
-struct tvr_msg_material_destroy
-{
-	int32_t mat_id;
-};
-
-/* Viewpoint */
-
-struct tvr_msg_pov_set_loc
-{
-	float x, y, z;
-};
-
-struct tvr_msg_pov_set_rot
-{
-	float x, y, z;
-};
-
-struct tvr_msg_pov_set_fov
-{
-	float fov;
-};
+/* Misc. requests */
 
 int recv_tvr_set_version(
 	int fd,
@@ -167,6 +56,11 @@ int recv_tvr_mesh_bind_mat(
 	struct tvr_msg_mesh_bind_mat *msg
 );
 
+int recv_tvr_mesh_bind_tex(
+	int fd,
+	struct tvr_msg_mesh_bind_tex *msg
+);
+
 int recv_tvr_mesh_set_loc(
 	int fd,
 	struct tvr_msg_mesh_set_loc *msg
@@ -192,6 +86,11 @@ int recv_tvr_mesh_destroy(
 int recv_tvr_texture_create(
 	int fd,
 	struct tvr_msg_texture_create *msg
+);
+
+int recv_tvr_texture_write(
+	int fd,
+	struct tvr_msg_texture_write *msg
 );
 
 int recv_tvr_texture_destroy(
@@ -225,6 +124,11 @@ int recv_tvr_pov_set_loc(
 int recv_tvr_pov_set_rot(
 	int fd,
 	struct tvr_msg_pov_set_rot *msg
+);
+
+int recv_tvr_pov_look_at(
+	int fd,
+	struct tvr_msg_pov_look_at *msg
 );
 
 int recv_tvr_pov_set_fov(