+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
-v0.1.2
+## v0.1.2
 + Set default server location to the environment variable `TVR_DISPLAY`
++ Added support for TVrender protocol version 2
++ More memory fixes
++ Arguments now pass down to child processes
++ AddressSanitizer now only applies to debug builds
 
-v0.1.1
+## v0.1.1
 + Memory error fixes
 + Added changelog
 + Updated readme